fix(player): resume favourites queue on system playback resumption
onPlaybackResumption was unimplemented, so a lock-screen/notification
Previous/Next press after the whole process died during sleep hit
Media3's default (a failed future) and had no queue to navigate.
Rebuilds it from the persisted last-played station and favourites
sort, same as loadStationPaused, but for when the app itself never
reopens. Shared the queue/sort helpers between MainViewModel and
PlayerService via data/StationQueue.kt instead of duplicating them.
